+++ packages/components/scripts/jobs/synergy/cem.js
@@ -3,12 +3,12 @@ import {
   formatFile,
   getPath,
   job,
-} from './shared.js';
+} from '../shared.js';
 
 /**
  * Run cem for the components package
  */
-export const runCem = job('Creating component manifest...', async (componentDistDir = 'dist') => {
+export const runCem = job('Synergy: Creating component manifest...', async (componentDistDir = 'dist') => {
   await execPromise(
     `cem analyze --litelement --outdir ${componentDistDir}`,
     { stdio: 'inherit' },