+++ packages/components/src/events/events.ts
@@ -13,6 +13,7 @@ export type * from './syn-burger-menu-open.ts';
 export type * from './syn-change.ts';
 export type * from './syn-clear.ts';
 export type * from './syn-close.ts';
+export type * from './syn-default-settings-changed.ts';
 export type * from './syn-error.ts';
 export type * from './syn-focus.ts';
 export type * from './syn-hide.ts';