+++ packages/components/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@synergy-design-system/components-v2.21.1](https://github.com/synergy-design-system/synergy-design-system/compare/components/2.21.0...components/2.21.1) (2025-02-13)
+
+
+### Bug Fixes
+
+* 🐛 Closable syn-alert closes on whole vertical area instead of the close icon only ([#774](https://github.com/synergy-design-system/synergy-design-system/issues/774)) ([3793ab2](https://github.com/synergy-design-system/synergy-design-system/commit/3793ab283cbb26bccc57612eb56bdc90aa3b563b))
+
+# [@synergy-design-system/components-v2.21.0](https://github.com/synergy-design-system/synergy-design-system/compare/components/2.20.1...components/2.21.0) (2025-02-12)
+
+
+### Features
+
+* ✨ Allow to set global defaults for Synergy Component props ([#733](https://github.com/synergy-design-system/synergy-design-system/issues/733)) ([d5cf6fb](https://github.com/synergy-design-system/synergy-design-system/commit/d5cf6fbb07a496f142c15e2f021b65a2561a4e3a))
+
 # [@synergy-design-system/components-v2.20.1](https://github.com/synergy-design-system/synergy-design-system/compare/components/2.20.0...components/2.20.1) (2025-02-10)